+++ src/main/java/pl/pkrysztofiak/gridview/view/View.java
@@ -2,11 +2,11 @@ package pl.pkrysztofiak.gridview.view;
 
 import javafx.scene.layout.BorderPane;
 import pl.pkrysztofiak.gridview.model.Model;
-import pl.pkrysztofiak.gridview.view.panels.PanelsView;
+import pl.pkrysztofiak.gridview.view.layout.grid.GridPanelsView;
 
 public class View extends BorderPane {
     
     public View(Model model) {
-        setCenter(new PanelsView(model.getPanels()));
+        setCenter(new GridPanelsView(model.getPanels()));
     }
 }
