+++ src/main/java/pl/pkrysztofiak/gridview/view/layout/grid/GridPanelsView.java
@@ -1,4 +1,4 @@
-package pl.pkrysztofiak.gridview.view.panels;
+package pl.pkrysztofiak.gridview.view.layout.grid;
 
 import java.util.concurrent.TimeUnit;
 
@@ -9,17 +9,17 @@ import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.layout.Pane;
-import pl.pkrysztofiak.gridview.model.panels.PanelsModel;
+import pl.pkrysztofiak.gridview.model.layout.grid.GridPanelsModel;
 import pl.pkrysztofiak.gridview.model.panels.grid.VGridLineModel;
 import pl.pkrysztofiak.gridview.view.panels.grid.VGridLineView;
 
-public class PanelsView extends Pane {
+public class GridPanelsView extends Pane {
 
     private final ObservableList<VGridLineView> vGridLines = FXCollections.observableArrayList();
     public final Observable<Double> widthObservable = JavaFxObservable.valuesOf(widthProperty()).map(Number::doubleValue);
     public final Observable<Double> heightObservable = JavaFxObservable.valuesOf(heightProperty()).map(Number::doubleValue);
     
-    public PanelsView(PanelsModel panelsModel) {
+    public GridPanelsView(GridPanelsModel panelsModel) {
         Bindings.bindContent(getChildren(), vGridLines);
 
         Observable.fromIterable(panelsModel.vGridLines.vGridLines).delay(0, TimeUnit.SECONDS, JavaFxScheduler.platform()).subscribe(this::onVGridLineModelAdded);
